Refactor code suggested by reviewer
