+++ book.rb
@@ -9,7 +9,7 @@ class Book
     @rentals = []
   end
 
-  def add_rental(date, person)
-    Rental.new(date, person, self)
+  def add_rental(rental)
+    @rentals << rental
   end
 end
